test release

+++ lib/database/include/Car.h
@@ -10,14 +10,14 @@ public:
     uint64_t car_id;
     std::string name;
     uint64_t price;
-    uint64_t consumption;
+    double consumption;
     uint64_t capacity;
     std::string fuel;
     std::string picture_path;
     std::string town;
     std::string color;
     std::string brand;
-    Car(uint64_t id, std::string name, uint64_t price, uint64_t consumption, uint64_t capacity,
+    Car(uint64_t id, std::string name, uint64_t price, double consumption, uint64_t capacity,
         std::string fuel, std::string picture_path, std::string town, std::string color, std::string brand);
 };
